requirements.txt
jupyter==1.0.0
numpy==1.24.3
pandas==2.0.2
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
streamlit==1.23.1
torch==2.0.1
transformers==4.30.1
lxml==4.9.3
datasets==2.14.5
langchain==0.0.299
black==23.9.1